// src/app/pages/items/models/item-types.ts
export enum ItemTypes {
  MCQ = 'MCQ',
  MRQ = 'MRQ',
  //ESSAY_PLAIN_TEXT = 'ESSAY_PLAIN_TEXT',
  ESSAY_RICH_TEXT = 'ESSAY_RICH_TEXT',
  CLOZE_TEXT = 'CLOZE_TEXT',
  CLOZE_DROPDOWN = 'CLOZE_DROPDOWN',
  SHORT_TEXT = 'SHORT_TEXT',
  TRUE_FALSE = 'TRUE_FALSE',
  YES_NO = 'YES_NO',
  ASSOCIATION = 'ASSOCIATION',
  CHOICE_MATRIX = 'CHOICE_MATRIX',
  ORDER_LIST = 'ORDER_LIST',
  CLOZE_TEXT_IMAGE = 'CLOZE_TEXT_IMAGE',
  CLOZE_DROPDOWN_IMAGE = 'CLOZE_DROPDOWN_IMAGE',
  IMAGE_DRAG_AND_DROP = 'IMAGE_DRAG_AND_DROP',
  ANY = "ANY"
}